requirements_dev.txt
pip==22.3
bump2version==0.5.11
wheel==0.33.6
watchdog==0.9.0
flake8==3.7.8
tox==3.14.0
coverage==4.5.4
Sphinx==1.8.5
twine==1.14.0
Pillow>=9.0
pysimplegui==4.60.4
pytesseract==0.3.10
ConfigArgParse==1.5.3
auto-py-to-exe==2.23.1
